wrapped up the fixes and features for this version

+++ green-screen/data-access-layer/runtrxs.ps1
@@ -1,18 +1,17 @@
-
-$WorkingDir = "C:\Users\ag4488\Documents\Visual Studio 2019\Projects\cobol-nodejs-data-hooks\cobol-os"
+$WorkingDir=$args[0]
+Set-Location $WorkingDir
 
 $CurrMasterDataFile = Join-Path $WorkingDir "\data\students.dat"
 $fileExist = Test-Path $CurrMasterDataFile
-if ($fileExist){
-    Remove-Item $CurrMasterDataFile;
-}
-
+if ($fileExist) {Remove-Item $CurrMasterDataFile}
 $OrigMasterDataFile = Join-Path $WorkingDir "\data\students_orig.dat"
 $DestinationFile = Join-Path $WorkingDir "\data\students.dat"
-    
 Copy-Item $OrigMasterDataFile -Destination $DestinationFile
 
 $Executable = Join-Path $WorkingDir "\studentwrite.exe"
 & $Executable
 
-    
+# remove students.dat file and rename students1.dat
+Remove-Item $CurrMasterDataFile
+$NewMasterDataFile = Join-Path $WorkingDir "\data\students1.dat"
+Rename-Item -Path $NewMasterDataFile -NewName "students.dat"
